Separate the test file based on target examples
