--- src/__tests__/stylelint-no-unused-selectors.spec.ts
@@ -1,90 +1,0 @@
-import path from 'path';
-import stylelint from 'stylelint';
-import { parseResult } from './__helpers__/parse-result';
-
-const fixturesRoot = path.join(__dirname, '..', '..', 'examples');
-const configFilePath = path.join(fixturesRoot, '.stylelintrc.js');
-
-test('Disabllow selectors that are not used in a HTML file', async (): Promise<
-  void
-> => {
-  const options = {
-    configFile: configFilePath,
-    files: path.join(fixturesRoot, 'html', '*.css'),
-  };
-
-  const result = await stylelint.lint(options);
-
-  expect(result.errored).toBe(true);
-  expect(parseResult(result)).toMatchSnapshot();
-});
-
-test('Disabllow selectors that are not used in a JSX file', async (): Promise<
-  void
-> => {
-  const options = {
-    configFile: configFilePath,
-    files: path.join(fixturesRoot, 'jsx', '*.css'),
-  };
-
-  const result = await stylelint.lint(options);
-
-  expect(result.errored).toBe(true);
-  expect(parseResult(result)).toMatchSnapshot();
-});
-
-test('Disabllow selectors that are not used in a flow-typed JSX file', async (): Promise<
-  void
-> => {
-  const options = {
-    configFile: configFilePath,
-    files: path.join(fixturesRoot, 'jsx-with-flow', '*.css'),
-  };
-
-  const result = await stylelint.lint(options);
-
-  expect(result.errored).toBe(true);
-  expect(parseResult(result)).toMatchSnapshot();
-});
-
-test('Disabllow selectors that are not used in a JSX file using CSS Modules', async (): Promise<
-  void
-> => {
-  const options = {
-    configFile: configFilePath,
-    files: path.join(fixturesRoot, 'jsx-with-css-modules', '*.css'),
-  };
-
-  const result = await stylelint.lint(options);
-
-  expect(result.errored).toBe(true);
-  expect(parseResult(result)).toMatchSnapshot();
-});
-
-test('Disabllow selectors that are not used in a TSX file', async (): Promise<
-  void
-> => {
-  const options = {
-    configFile: configFilePath,
-    files: path.join(fixturesRoot, 'tsx', '*.css'),
-  };
-
-  const result = await stylelint.lint(options);
-
-  expect(result.errored).toBe(true);
-  expect(parseResult(result)).toMatchSnapshot();
-});
-
-test('Disabllow selectors that are not used in a TSX file using CSS Modules', async (): Promise<
-  void
-> => {
-  const options = {
-    configFile: configFilePath,
-    files: path.join(fixturesRoot, 'tsx-with-css-modules', '*.css'),
-  };
-
-  const result = await stylelint.lint(options);
-
-  expect(result.errored).toBe(true);
-  expect(parseResult(result)).toMatchSnapshot();
-});
